requirements.txt
trimesh
numpy
scipy
scikit-learn
rtree
pandas
PIL
tqdm
pyyaml